Switch to using Groovy 4
GORM doesn't support Groovy 4 yet, so I removed them from settings.gradle
and update the documentation to explain why they are missing.

Switched to using the imported catalog as much as possible.

I need to go through the dependencies, as I have a feeling some of them
are vestigial, and need pruning for the new build plugins